refactor: move ContainerTarget from cella-docker to cella-backend
ContainerTarget::resolve() now takes &dyn ContainerBackend instead of
&DockerClient. The label-based search is reimplemented using
list_cella_containers() + in-memory filter (no bollard dependency).

CLI commands updated to import from cella_backend. cella-docker
re-exports for backward compatibility.

Ref: #15
