+++ crates/cella-backend/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod error;
 pub mod lifecycle;
 pub mod names;
+pub mod resolve;
 pub mod traits;
 pub mod types;
 
@@ -12,6 +13,7 @@ pub use names::{
     compose_labels, compose_project_name, container_labels, container_name, image_name,
     image_name_with_features, worktree_labels,
 };
+pub use resolve::ContainerTarget;
 pub use traits::{BoxFuture, ComposeBackend, ContainerBackend, Platform};
 pub use types::{
     BackendKind, BuildOptions, ContainerInfo, ContainerState, CreateContainerOptions, DeviceSpec,
