+++ crates/cella-cli/src/commands/down.rs
@@ -8,7 +8,8 @@ use super::up::OutputFormat;
 use cella_compose::discovery;
 use cella_daemon::daemon;
 use cella_daemon::shared::running_cella_container_count;
-use cella_docker::{ContainerInfo, ContainerState, ContainerTarget};
+use cella_backend::ContainerTarget;
+use cella_docker::{ContainerInfo, ContainerState};
 use cella_env::paths::{cella_data_dir, daemon_socket_path};
 
 use crate::picker;
@@ -94,9 +95,8 @@ impl DownArgs {
 
     pub async fn execute(
         self,
-        backend: Option<&crate::backend::BackendChoice>,
+        _backend: Option<&crate::backend::BackendChoice>,
     ) -> Result<(), Box<dyn std::error::Error>> {
-        let _ = backend; // TODO: use resolve_backend once ContainerTarget is migrated
         let client = super::connect_docker(self.docker_host.as_deref())?;
 
         let workspace_folder = if let Some(ref branch_name) = self.branch {
