+++ crates/cella-cli/src/commands/shell.rs
@@ -3,7 +3,8 @@ use std::path::PathBuf;
 use clap::Args;
 use tracing::debug;
 
-use cella_docker::{ContainerTarget, InteractiveExecOptions};
+use cella_backend::ContainerTarget;
+use cella_docker::InteractiveExecOptions;
 
 use crate::picker;
 
@@ -38,9 +39,8 @@ pub struct ShellArgs {
 impl ShellArgs {
     pub async fn execute(
         self,
-        backend: Option<&crate::backend::BackendChoice>,
+        _backend: Option<&crate::backend::BackendChoice>,
     ) -> Result<(), Box<dyn std::error::Error>> {
-        let _ = backend; // TODO: use resolve_backend once ContainerTarget is migrated
         let client = super::connect_docker(self.docker_host.as_deref())?;
 
         let target = ContainerTarget {
