+++ crates/cella-cli/src/commands/switch.rs
@@ -1,7 +1,8 @@
 use clap::Args;
 use tracing::debug;
 
-use cella_docker::{ContainerTarget, InteractiveExecOptions};
+use cella_backend::ContainerTarget;
+use cella_docker::InteractiveExecOptions;
 
 use crate::picker;
 
@@ -23,9 +24,8 @@ pub struct SwitchArgs {
 impl SwitchArgs {
     pub async fn execute(
         self,
-        backend: Option<&crate::backend::BackendChoice>,
+        _backend: Option<&crate::backend::BackendChoice>,
     ) -> Result<(), Box<dyn std::error::Error>> {
-        let _ = backend; // TODO: use resolve_backend once ContainerTarget is migrated
         let client = super::connect_docker(self.docker_host.as_deref())?;
 
         // Discover repo and list worktrees
